[#22] 개선
