+++ Projects/Shared/FirebaseShared/Interface/FirebaseServiceProtocol.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public protocol FirebaseServiceProtocol {
+  func configure()
+  func isConfigured() -> Bool
+}
